Add files via upload
PROJECT.c is basically the main project file I have built. The project requires two input files for checking plagiarism in between them...student1.c is basically the work of a student A for instance. then there's one student who only plays with the comment to hide plagiarism, another one plays with number of lines, some one else plays with the spaces in between lines! and someone even changes the sequence of lines or the sequence of functions defined etc. Then at the end there's an intelligent guy which uses all of the above techniques but still can't escape Brute Plagiarism detector. Go give it a try yourself, I can guarantee that you will enjoy it. This is my first-semester project in the Programming Fundamentals course in a 4 years degree program  of BSCS.